gradefactory/requirements.txt
google-generativeai
python-dotenv
fpdf
PyPDF2
pyinstaller
PyMuPDF
google-cloud-vision
pypdfium2
Pillow
requests